% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_boxplot.R
\name{resid_boxplot}
\alias{resid_boxplot}
\title{Boxplot of Residuals.}
\usage{
resid_boxplot(model)
}
\arguments{
\item{model}{Model fit using either lm, glm, lmer, or glmer.}
}
\value{
A boxplot of the residuals from the \code{model}.
}
\description{
Creates a boxplot on the residuals from a model.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_boxplot(model)

}
