% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_cookd.R
\name{resid_cookd}
\alias{resid_cookd}
\title{Cook's D Plot.}
\usage{
resid_cookd(model)
}
\arguments{
\item{model}{Model fit using either lm or glm.}
}
\value{
A plot of the Cook's D values versus observation numbers from the \code{model}.
The Horizontal line represents a cut-off to identify highly influential points. The horizontal line is
placed at 4/n where n is the number of data points used in the \code{model}.
}
\description{
Creates a Cook's D plot with the Cook's D values versus the observation number.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_cookd(model)
}
