% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_hist.R
\name{resid_hist}
\alias{resid_hist}
\title{Histogram of Residuals.}
\usage{
resid_hist(model, bins = NA)
}
\arguments{
\item{model}{Model fit using either lm, glm, lmer, or glmer.}
}
\value{
A histogram of the residuals from the \code{model} with a normal
density curve overlaid with mean equal to the mean of the residuals and
standard deviation equal to the standard deviation of the residuals.
}
\description{
Creates a histogram of the residuals from a model.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_hist(model)
}
