% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_interact.R
\name{resid_interact}
\alias{resid_interact}
\title{Interaction Versions of Residual Diagnostics Plots.}
\usage{
resid_interact(model, plot = NA, bins = NA, type = NA, smoother = FALSE,
  theme = "bw", axis.text.size = 10, title.text.size = 12,
  title.opt = TRUE, qqline = TRUE)
}
\arguments{
\item{model}{Model fit using either \code{lm}, \code{glm}, \code{lmer}, or \code{glmer}.}

\item{plot}{Plot chosen to create interactive plot. (See details for options.)}

\item{bins}{Number of bins for histogram of the residuals.}

\item{type}{The user may specify a type of residuals to use. Otherwise, the default
residual type for each model is used. (See details for options.)}

\item{smoother}{Indicates whether or not to include a smoother on the residual plot.
Specify TRUE or FALSE. Default is set to FALSE.}

\item{theme}{ggplot2 theme to be used. Options are \code{"bw"}, \code{"classic"}, and
\code{"grey"} (or \code{"gray"}). Default is \code{"bw"}.}

\item{axis.text.size}{Specifies the size of the text for the axis labels of the plot.}

\item{title.text.size}{Specifies the size of the text for the titles of the plot.}

\item{title.opt}{Indicates whether or not to include a title on the plot.
Specify TRUE or FALSE. Default is set to TRUE.}
}
\value{
The interactive residual diagnostic plot specified.
}
\description{
Creates interaction versions of all residual diagnostic plots.
}
\examples{
# Fit a linear regression model and plot the residuals using the default panel
lm_model <- lm(Volume ~ Girth, data = trees)
resid_interact(lm_model, bins = 30)
}
