% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_label.R
\name{resid_label}
\alias{resid_label}
\title{Labels for Plots.}
\usage{
resid_label(type = NA, model)
}
\arguments{
\item{type}{Specifies type of desired residuals. If NA, the default residual type for each  model is used.}

\item{model}{Model fit using either lm, glm, lmer, or glmer.}
}
\value{
The appropriate residual label.
}
\description{
Creates a label for the plots based on the type of residuals used.
}
