% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_lev.R
\name{resid_lev}
\alias{resid_lev}
\title{Residual-Leverage plot.}
\usage{
resid_lev(model)
}
\arguments{
\item{model}{Model fit using lm.}
}
\value{
A plot of residuals versus leverage values from the \code{model}.
}
\description{
Creates a plot of the residuals versus leverage from a model.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
residlev_plot(model)
}
