% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_ls.R
\name{resid_ls}
\alias{resid_ls}
\title{Scale-Location Plot.}
\usage{
resid_ls(model)
}
\arguments{
\item{model}{Model fit using either lm or glm.}
}
\value{
A plot of the square roote of the standardized residuals versus predicted values from the \code{model}
 with smooth curve fit using "loess".
}
\description{
Creates a scale-location plot with the square root of the standardized residuals versus predicted values from a model.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_ls(model)
}
