% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_plot.R
\name{resid_plot}
\alias{resid_plot}
\title{Residual Plot.}
\usage{
resid_plot(resid, pred)
}
\arguments{
\item{resid}{Residuals from a model.}

\item{pred}{Predicted values from a model.}
}
\value{
A residual plot of \code{resid} versus \code{pred} with a horizontal line
with an intercept of 0.
}
\description{
Creates a residual plot with residuals versus predicted values from a model.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_plot(model$residuals, model$fitted.values)
}

