% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_qq.R
\name{resid_qq}
\alias{resid_qq}
\title{Q-Q Plot.}
\usage{
resid_qq(resid)
}
\arguments{
\item{resid}{Residuals from a model.}
}
\value{
A Q-Q Plot of \code{resid}. The method for creating this Q-Q plot follows that
used by SAS:

\url{http://support.sas.com/documentation/cdl/en/sgug/59902/HTML/default/viewer.htm#fit_sect51.htm#sg_fit_fitresidualnormalquantiles}

After sorting the residuals in ascending order, for each of the ith ordered residuals,
the following quantile is computed:

\deqn{\Phi^{-1} * ((i - 0.375) / (n + 0.25))}

Each pair of points is then plotted to create the Q-Q plot. The line drawn on the
plot has an intercept equal to the mean of the residuals and a slope equal to the
standard deviation of the residuals. Data that is approximately normal will fall
very close to or on the line.
}
\description{
Creates a Q-Q plot on the residuals from a model.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_qq(model$residuals)
}

