% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_auxpanel.R
\name{resid_auxpanel}
\alias{resid_auxpanel}
\title{Panel of Diagnostic Residual Plots.}
\usage{
resid_auxpanel(resid, pred, plots = "SAS", bins = NA, scale = 1,
  smoother = FALSE, theme = "bw", axis.text.size = 10,
  title.text.size = 12, title.opt = TRUE, qqline = TRUE,
  qqbands = FALSE, ind.ncol = 2)
}
\arguments{
\item{resid}{The residuals from the model.}

\item{pred}{The fitted values from the model.}

\item{plots}{Plots chosen to include in the panel of plots. (See details for options.)}

\item{bins}{Number of bins for histogram of the residuals.}

\item{scale}{Scale of graphs in panel. Takes values in (0,1].}

\item{smoother}{Indicates whether or not to include a smoother on the residual plot.
Specify TRUE or FALSE. Default is set to FALSE.}

\item{theme}{ggplot2 theme to be used. Options are \code{"bw"}, \code{"classic"}, and
\code{"grey"} (or \code{"gray"}). Default is \code{"bw"}.}

\item{axis.text.size}{Specifies the size of the text for the axis labels of all plots.}

\item{title.text.size}{Specifies the size of the text for the titles of all plots.}

\item{title.opt}{Indicates whether or not to include a title on the plots.
Specify TRUE or FALSE. Default is set to TRUE.}

\item{ind.ncol}{Sets the number of columns in the panel when more than one individual plot
has been specified. Default is set to 2 columns.}
}
\value{
A panel of residual diagnostic plots containing plots specified.
}
\description{
Creates a panel of residual diagnostic plots given inputs of residuals and fitted values.
}
\details{
The following grid options can be chosen for the \code{plots} argument.
\itemize{
  \item "SAS": This is the default option. It creates a panel of a residual plot,
  a normal quantile plot of the residuals, a histogram of the residuals, and a
  boxplot of the residuals. This was modeled after the residpanel option in proc
  mixed from SAS version 9.4.
  \item A vector of individual plots can also be specified. For example, one can
  specify \code{plots = c("boxplot", "hist")} or \code{plots = "qq"}. The individual
  plot options are as follows.
  \itemize{
    \item \code{"boxplot"}: A boxplot of residuals.
    \item \code{"hist"}: A histogram of residuals.
    \item \code{"qq"}: A normal quantile plot of residuals.
    \item \code{"residplot"}: A plot of residuals versus predicted values.
  }
}
}
\examples{
# Fit a linear regression model and plot the residuals using the default panel
lm_model <- lm(Volume ~ Girth, data = trees)
resid_spanel(resid(lm_model), fitted(lm_model), bins = 30)

# Fit a generalized linear regression model and plot the residuals using
# the default panel
glm_model <- glm(Height ~ Girth, family = "poisson", data = trees)
resid_spanel(resid(glm_model), fitted(glm_model), bins = 30)

# Generate normal data, fit a mixed effects model, and plot the residuals
# using the default panel
library(lme4)
d1 <- data.frame(y = rnorm(54, 20, 4), trt = rep(c("A", "B"), each = 27), subject = rep(1:18, each = 3))
lmer_model <- lmer(y ~ trt + (1|subject), data = d1)
resid_spanel(resid(lmer_model), fitted(lmer_model), bins = 30)

# Generate Poisson data, fit a mixed effects model, and plot the residuals
# using the default panel
d2 <- data.frame(y = rpois(54, 3), trt = rep(c("A", "B"), each = 27), subject = rep(1:18, each = 3))
glmer_model <- glmer(y ~ trt + (1|subject), family = "poisson", data = d2)
resid_spanel(resid(glmer_model), fitted(glmer_model), bins = 30)
}
