% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_auxpanel.R
\name{resid_auxpanel}
\alias{resid_auxpanel}
\title{Panel of Diagnostic Residual Plots.}
\usage{
resid_auxpanel(residuals, predicted, plots = "SAS", bins = NA,
  smoother = FALSE, qqline = TRUE, qqbands = FALSE, scale = 1,
  theme = "bw", axis.text.size = 10, title.text.size = 12,
  title.opt = TRUE, ind.ncol = 2)
}
\arguments{
\item{plots}{Plots chosen to include in the panel of plots. (See details for options.)}

\item{bins}{Number of bins for histogram of the residuals.}

\item{smoother}{Indicates whether or not to include a smoother on the residual plot.
Specify TRUE or FALSE. Default is set to FALSE.}

\item{qqline}{Indicates whether to include a 1-1 line on the qq-plot. Specify TRUE or
FALSE. Default is set to TRUE.}

\item{qqbands}{Indicates whether to include confidence bands on the qq-plot. Specify
TRUE or FALSE. Default is set to FALSE.}

\item{scale}{Scales the size of the graphs in a panel. Takes values in (0,1].}

\item{theme}{ggplot2 theme to be used. Options are \code{"bw"}, \code{"classic"}, and
\code{"grey"} (or \code{"gray"}). Default is \code{"bw"}.}

\item{axis.text.size}{Specifies the size of the text for the axis labels of all plots.}

\item{title.text.size}{Specifies the size of the text for the titles of all plots.}

\item{title.opt}{Indicates whether or not to include a title on the plots.
Specify TRUE or FALSE. Default is set to TRUE.}

\item{ind.ncol}{Sets the number of columns in the panel when more than one individual plot
has been specified. Default is set to 2 columns.}

\item{resid}{The residuals from the model.}

\item{pred}{The fitted values from the model.}
}
\value{
A panel of residual diagnostic plots containing plots specified.
}
\description{
Creates a panel of residual diagnostic plots given inputs of residuals and fitted values.
}
\details{
The following grid options can be chosen for the \code{plots} argument.
\itemize{
  \item "SAS": This is the default option. It creates a panel of a residual plot,
  a normal quantile plot of the residuals, a histogram of the residuals, and a
  boxplot of the residuals. This was modeled after the residpanel option in proc
  mixed from SAS version 9.4.
  \item A vector of individual plots can also be specified. For example, one can
  specify \code{plots = c("boxplot", "hist")} or \code{plots = "qq"}. The individual
  plot options are as follows.
  \itemize{
    \item \code{"boxplot"}: A boxplot of residuals
    \item \code{"hist"}: A histogram of residuals
    \item \code{"qq"}: A normal quantile plot of residuals
    \item \code{"resid"}: A plot of residuals versus predicted values
  }
}

Details on the creation of the plots can be found in the details section of the help
file for \code{resid_panel}.
}
\examples{
# Fit a linear regression model to predict the volume of a tree based on the
# girth of a tree using the R "trees" data
lm_model1 <- lm(Volume ~ Girth, data = trees)

# Plot the residuals using the default panel
resid_auxpanel(resid(lm_model1), fitted(lm_model1), bins = 30)

# Fit a linear model to compare the weights of plants bewteen different
# treatment groups using the R "PlantGrowth" data
lm_model2 <- lm(weight ~ group, data = PlantGrowth)

# Create a grid of the residual plot and qq-plot
resid_auxpanel(resid(lm_model2), fitted(lm_model2), plots = c("resid", "qq"))

# Fit a generalized linear regression model using a Poisson family to compare
# the insect counts between different sprays from the R "InsectSprays" data
glm_model <- glm(count ~ spray, family = "poisson", data = InsectSprays)

# Create the SAS panel of plots without titles, with a grey theme, and with a
# smaller scaling of the plots
resid_auxpanel(resid(glm_model), fitted(glm_model), title.opt = F, theme = "grey",
               bins = 30, scale = 0.9)
}
