% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_hist.R
\name{resid_hist}
\alias{resid_hist}
\title{Histogram of Residuals.}
\usage{
resid_hist(resid, bins = NA)
}
\arguments{
\item{resid}{Residuals from a model.}
}
\value{
A histogram of \code{resid} with a normal density curve overlaid with
mean equal to the mean of the residuals and standard deviation equal to the
standard deviation of the residuals.
}
\description{
Creates a histogram of the residuals from a model.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_hist(model$residuals)
}

