% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_panel.R
\name{resid_panel}
\alias{resid_panel}
\title{Panel of Diagnostic Residual Plots.}
\usage{
resid_panel(resid, pred, bins = NA)
}
\arguments{
\item{resid}{Residuals from a model.}

\item{pred}{Predicted values from a model.}
}
\value{
A panel of residual diagnostic plots including a residual plot, a histogram of the
residuals, a qqplot of the residuals, and a boxplot of the residuals.
}
\description{
Creates a panel of residual diagnostic plots.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_panel(model$residuals, model$fitted.values)

}

