% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_panel.R
\name{resid_panel}
\alias{resid_panel}
\title{Panel of Diagnostic Residual Plots.}
\usage{
resid_panel(model, plots = "SAS", bins = NA, scale = 1)
}
\arguments{
\item{model}{Model fit using either lm or glm.}

\item{plots}{Plots chosen to include in the panel of plots. (See details for options.)}

\item{bins}{Number of bins for histogram of the residuals.}

\item{scale}{Scale of graphs in panel. Takes values in (0,1].}
}
\value{
A panel of residual diagnostic plots containing plots specified.
}
\description{
Creates a panel of residual diagnostic plots.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_panel(model)
}
