% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_panel.R
\name{resid_panel}
\alias{resid_panel}
\title{Panel of Diagnostic Residual Plots.}
\usage{
resid_panel(model, plots = "SAS", bins = NA, scale = 1, type = NA,
  smoother = NA, theme = NA)
}
\arguments{
\item{model}{Model fit using either \code{lm}, \code{glm}, \code{lmer}, or \code{glmer}.}

\item{plots}{Plots chosen to include in the panel of plots. (See details for options.)}

\item{bins}{Number of bins for histogram of the residuals.}

\item{scale}{Scale of graphs in panel. Takes values in (0,1].}
}
\value{
A panel of residual diagnostic plots containing plots specified.
}
\description{
Creates a panel of residual diagnostic plots.
}
\details{
The following grid options can be chosen for the \code{plots} argument.
\itemize{
  \item "all": This creates a panel of all plot types included in the package.
  \item "R": This creates a panel of a residual plot, a normal quantile plot of
  the residuals, and a leverage versus residuals plot. This was modeled after the
  plots shown in R if the \code{plots()} base function is applied to a \code{lm}
  model.
  \item "SAS": This is the default option. It creates a panel of a residual plot,
  a normal quantile plot of the residuals, a histogram of the residuals, and a
  boxplot of the residuals. This was modeled after the residpanel option in proc
  mixed from SAS version 9.4.
  \item A vector of individual plots can also be specified. For example, one can
  specify \code{plots = c("boxplot", "hist")} or \code{plots = "qq"}. The individual
  plot options are as follows.
  \itemize{
    \item "boxplot": A boxplot of residuals.
    \item "hist": A histogram of residuals.
    \item "ls": A location scale plot of the residuals.
    \item "qq": A normal quantile plot of residuals.
    \item "residlev": A plot of leverage values versus residuals.
    \item "residplot": A plot of residuals versus predicted values.
  }
}
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_panel(model)
}
