% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_panel.R
\name{resid_panel}
\alias{resid_panel}
\title{Panel of Diagnostic Residual Plots.}
\usage{
resid_panel(model, plots = "SAS", type = NA, bins = NA,
  smoother = FALSE, qqline = TRUE, qqbands = FALSE, scale = 1,
  theme = "bw", axis.text.size = 10, title.text.size = 12,
  title.opt = TRUE, ind.ncol = 2)
}
\arguments{
\item{model}{Model fit using either \code{lm}, \code{glm}, \code{lmer}, or \code{glmer}.}

\item{plots}{Plots chosen to include in the panel of plots. Default is set to "SAS".
(See details for options.)}

\item{type}{The user may specify a type of residuals to use. Otherwise, the default
residual type for each model is used. (See details for options.)}

\item{bins}{Number of bins to use when creating a histogram of the residuals.}

\item{smoother}{Indicates whether or not to include a smoother on the residual plot.
Specify TRUE or FALSE. Default is set to FALSE.}

\item{qqline}{Indicates whether to include a 1-1 line on the qq-plot. Specify TRUE or
FALSE. Default is set to TRUE.}

\item{qqbands}{Indicates whether to include confidence bands on the qq-plot. Specify
TRUE or FALSE. Default is set to FALSE.}

\item{scale}{Scales the size of the graphs in a panel. Takes values in (0,1].}

\item{theme}{ggplot2 theme to be used. Current options are \code{"bw"}, \code{"classic"},
and \code{"grey"} (or \code{"gray"}). Default is \code{"bw"}.}

\item{axis.text.size}{Specifies the size of the text for the axis labels of all plots
in the panel.}

\item{title.text.size}{Specifies the size of the text for the titles of all plots in
the panel.}

\item{title.opt}{Indicates whether or not to include a title on the plots in the panel.
Specify TRUE or FALSE. Default is set to TRUE.}

\item{ind.ncol}{When a vector of individual plots has been requested, this specifies the
number of columns in the panel. Default is set to 2 columns.}
}
\value{
A panel of residual diagnostic plots containing plots specified.
}
\description{
Creates a panel of residual diagnostic plots given a model. Currently accepts models of
type "lm", "glm", "lmerMod", and "glmerMod".
}
\details{
The first two sections below contain information on the available input options for the
\code{plots} and \code{type} arguments in \code{resid_panel}. The third section contains
information on the details relating to the creation of the plots.

\strong{Options for} \code{plots}

The following options can be chosen for the \code{plots} argument.
\itemize{
  \item "all": This creates a panel of all plot types included in the package that
  are available for the model type input into \code{residpanel}. (See note below.)
  \item "R": This creates a panel with a residual plot, a normal quantile plot of
  the residuals, and a leverage versus residuals plot. This was modeled after the
  plots shown in R if the \code{plots()} base function is applied to an \code{lm}
  model. This option can only be used with an \code{lm} or \code{glm} model.
  \item "SAS": This is the default option. It creates a panel with a residual plot,
  a normal quantile plot of the residuals, a histogram of the residuals, and a
  boxplot of the residuals. This was modeled after the residpanel option in proc
  mixed from SAS version 9.4.
  \item A vector of individual plots can also be specified. For example, one can
  specify \code{plots = c("boxplot", "hist")} or \code{plots = "qq"}. The individual
  plot options are as follows.
  \itemize{
    \item \code{"boxplot"}: A boxplot of residuals.
    \item \code{"cookd"}: A plot of Cook's D values versus observation number.
    \item \code{"hist"}: A histogram of residuals.
    \item \code{"ls"}: A location scale plot of the residuals.
    \item \code{"qq"}: A normal quantile plot of residuals.
    \item \code{"residlev"}: A plot of leverage values versus residuals.
    \item \code{"residplot"}: A plot of residuals versus predicted values.
    \item \code{"respred":}: A plot of the response variable versus the predicted values.
  }
}
Note: \code{"cookd"}, \code{"ls"}, and \code{"residlev"} are not available for "lmer"
and "glmer" models.

\strong{Options for} \code{type}

Several residual types are available to be requested based on the model type that
is input into \code{resid_panel}. These currently are as follows.
\itemize{
  \item \code{lm} residual options
  \itemize{
    \item \code{"pearson"}: The Pearson residuals.
    \item \code{"response"}: The raw residuals. (Default for "lm".)
    \item \code{"standardized"}: The standardized raw residuals.
  }
  \item \code{glm} residual options
  \itemize{
    \item \code{"deviance"}: The deviance residuals. (Default for "glm".)
    \item \code{"pearson"}: The Pearson residuals.
    \item \code{"response"}: The raw residuals.
    \item \code{"stand.deviance"}: The standardized deviance residuals.
    \item \code{"stand.pearson"}: The standardized Pearson residuals.
  }
  \item \code{lmer} residual options
  \itemize{
    \item \code{"pearson"}: The Pearson residuals. (Default for "lmer".)
    \item \code{"response"}: The raw residuals.
  }
  \item \code{glmer} residual options
  \itemize{
    \item \code{"deviance"}: The deviance residuals. (Default for "glmer".)
    \item \code{"pearson"}: The Pearson residuals.
    \item \code{"response"}: The raw residuals.
  }
}
Note: The plots of \code{"ls"} and \code{"lev"} only accept standarized residuals.

\strong{Details on the Creation of Plots}

Boxplot (\code{boxplot}): Boxplot of the residuals.

Cook's D (\code{cookd}): The horizontal line represents a cut-off to identify highly
influential points. The horizontal line is placed at 4/n where n is the number of
data points used in the \code{model}.

Histogram (\code{hist}): Plots a historgram of the residuals. The density
curve overlaid has mean equal to the mean of the residuals and standard deviation
equal to the standard deviation of the residuals.

Leverage Plot (\code{lev}): Plots the standardized residuals on the y-axis and the
leverage values on the x-axis. A lowess curve is overlaid, and Cook's D contours
are included for \eqn{\alpha = 0.5} and \eqn{\alpha = 1}.

Location-Scale Plot (\code{ls}): Plots the square root of the absolute value of the
standardized residuals on the y-axis and the predicted values on the x-axis. The
predicted values are plotted on the original scale for \code{glm} and \code{glmer}
models. A lowess curve is overlaid.

QQ Plot (\code{qq}): Makes use of the \code{R} package \code{qqplotr} for creating
a normal quantile plot of the residuals.

Residual Plot (\code{residplot}): Plots the residuals on the y-axis and the predicted
values on the x-axis. The predicted values are plotted on the original scale for
\code{glm} and \code{glmer} models.

Response vs. Predicted (\code{respred}): Plots the response variable from the model
on the y-axis and the predicted values on the x-axis. Both response variable and
predicted values are plotted on the original scale for \code{glm} and \code{glmer}
models.
}
\examples{
## --------------------------------------------------------------------------------
## Linear Regression Models
## --------------------------------------------------------------------------------

# Fit a linear regression model to predict the volume of a tree based on the
# girth of a tree using the R "trees" data
lm_model1 <- lm(Volume ~ Girth, data = trees)

# Plot the residuals using the default panel
resid_panel(lm_model1, bins = 30)

# Create a panel of all diagnostic plot options and add a smoother to the
# residual plot
resid_panel(lm_model1, bins = 30, plots = "all", smoother = TRUE)

# Fit a linear model to compare the weights of plants bewteen different
# treatment groups using the R "PlantGrowth" data
lm_model2 <- lm(weight ~ group, data = PlantGrowth)

# Create a panel of the R diagnostic plots for "lm" models using the classic
# theme
resid_panel(lm_model2, theme = "classic", plots = "R")

## --------------------------------------------------------------------------------
## Generalized Linear Regression Models
## --------------------------------------------------------------------------------

# Fit a generalized linear regression model using a Poisson family to compare
# the insect counts between different sprays from the R "InsectSprays" data
glm_model <- glm(count ~ spray, family = "poisson", data = InsectSprays)

# Plot the residuals using the default panel without titles and with a gray theme
resid_panel(glm_model, bins = 30, title.opt = FALSE, theme = "gray")
resid_panel(glm_model, bins = 30, title.opt = FALSE, theme = "grey")

# Plot the residuals using the default panel with standarized deviance residuals
resid_panel(glm_model, type = "stand.deviance")

## --------------------------------------------------------------------------------
## Linear Mixed Effects Model
## --------------------------------------------------------------------------------

# Load the lme4 package
library(lme4)

# Fit a linear mixed effect model to compare weights of chicks between diets using
# the R "ChickWeight" data and including chick as a random effect to account for the
# multiple measurements over time
lmer_model <- lmer(weight ~ Time + Diet + Time*Diet + (1|Chick), data = ChickWeight)

# Create a panel of the residual plot and the normal quantile plot with
# confidence bands
resid_panel(lmer_model, plots = c("residplot", "qq"), qqbands = TRUE)

## --------------------------------------------------------------------------------
## Generalized Linear Mixed Effects Model
## --------------------------------------------------------------------------------

# Generate Poisson data
example_data <- data.frame(y = rpois(54, 3),
                           trt = rep(c("A", "B"), each = 27),
                           subject = rep(1:18, each = 3))

# Fit a generalized linear mixed effects model with a Poisson family to compare
# the response between the treatments with a random effect for subject to
# account for the dependence within a subject
glmer_model <- glmer(y ~ trt + (1|subject), family = "poisson", data = example_data)

# Plot the residual plot with the size of the title and axis lables increased
resid_panel(glmer_model, plots = "residplot", title.text.size = 14, axis.text.size = 12)

# Plot the residual plot using the Pearson residuals
resid_panel(glmer_model, plots = "residplot", type = "pearson")
}
