% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_plot.R
\name{resid_plot}
\alias{resid_plot}
\title{Residual Plot.}
\usage{
resid_plot(model)
}
\arguments{
\item{model}{Model fit using either lm, glm, lmer, or glmer.}
}
\value{
A plot of the residuals versus predicted values from the \code{model}
 with a horizontal line through 0.
}
\description{
Creates a residual plot with residuals versus predicted values from a model.
}
\examples{
model <- lm(Volume ~ Girth, data = trees)
resid_plot(model)
}
